Removed majority of imports from main utils init and fixed issues
