+++ lightworks/qubit/gates/two_qubit_gates.py
@@ -21,7 +21,9 @@ lightworks.
 import numpy as np
 
 from lightworks.sdk.circuit import PhotonicCircuit, Unitary
-from lightworks.sdk.utils import permutation_mat_from_swaps_dict
+from lightworks.sdk.utils.permutation_conversion import (
+    permutation_mat_from_swaps_dict,
+)
 
 from .single_qubit_gates import H
