+++ lightworks/sdk/circuit/photonic_compiler.py
@@ -18,7 +18,7 @@ from typing import Any
 import numpy as np
 from numpy.typing import NDArray
 
-from lightworks.sdk.utils import ModeRangeError
+from lightworks.sdk.utils.exceptions import ModeRangeError
 
 from .photonic_components import Barrier, Component, Group, Loss
