+++ lightworks/sdk/circuit/photonic_components.py
@@ -20,7 +20,10 @@ from typing import Any
 import numpy as np
 from numpy.typing import NDArray
 
-from lightworks.sdk.utils import check_unitary, permutation_mat_from_swaps_dict
+from lightworks.sdk.utils.matrix import check_unitary
+from lightworks.sdk.utils.permutation_conversion import (
+    permutation_mat_from_swaps_dict,
+)
 
 from .parameterized_unitary import ParameterizedUnitary
 from .parameters import Parameter
