+++ lightworks/sdk/utils/__init__.py
@@ -13,20 +13,8 @@
 # limitations under the License.
 
 from .conversion import convert
-from .exceptions import *
-from .heralding_utils import add_heralds_to_state, remove_heralds_from_state
-from .matrix_utils import check_unitary
-from .permutation_conversion import permutation_mat_from_swaps_dict
-from .post_selection import (
-    DefaultPostSelection,
-    PostSelection,
-    PostSelectionFunction,
-    PostSelectionType,
-)
-from .post_selection_processing import process_post_selection
-from .random_utils import (
-    process_random_seed,
+from .post_selection import PostSelection, PostSelectionFunction
+from .random import (
     random_permutation,
     random_unitary,
 )
-from .task_utils import validate_states
