+++ lightworks/sdk/utils/heralding.py
