+++ lightworks/sdk/utils/random.py
@@ -1,4 +1,4 @@
-# Copyright 2024 Aegiq Ltd.
+# Copyright 2024 Aegiq Ltd.  # noqa: A005
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
