+++ lightworks/sdk/visualisation/draw_circuit_svg.py
@@ -27,7 +27,7 @@ from lightworks.sdk.circuit.photonic_components import (
     PhaseShifter,
     UnitaryMatrix,
 )
-from lightworks.sdk.utils import DisplayError
+from lightworks.sdk.utils.exceptions import DisplayError
 
 from .display_components_svg import DrawSVGComponents
 from .display_utils import process_parameter_value
